clean code

+++ src/main/java/com/icthh/xm/uaa/domain/properties/TenantProperties.java
@@ -82,8 +82,8 @@ public class TenantProperties {
 
         @Data
         public static class PasswordSettings {
-            private byte minLength = 0;
-            private byte maxLength = Byte.MAX_VALUE;
+            private int minLength = 0;
+            private int maxLength = Byte.MAX_VALUE;
             private String pattern;
             private String patternMessage;
             private boolean enableBackEndValidation = false;
